% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_landcover_desc.R
\name{add_landcover_desc}
\alias{add_landcover_desc}
\title{Add 'extra' ecosystem descriptions}
\usage{
add_landcover_desc(
  eco_desc,
  add_eco,
  clust_col = "cluster",
  add_clust_col = "cluster",
  add_name = "Landcover",
  colour_map = NULL,
  add_colour_col = "lc_col"
)
}
\arguments{
\item{eco_desc}{Dataframe of existing (floristic) ecosystem descriptions.}

\item{add_eco}{Dataframe of ecosystems to add.}

\item{clust_col}{Character name of column with cluster membership in
\code{eco_desc}.}

\item{add_clust_col}{Character name of column with cluster membership in
\code{add_eco}.}

\item{add_name}{Character name of any prefix to use in descriptions for extra
ecosystems. e.g. "Landcover" to make, say, "cropping" into "Landcover:
cropping".}

\item{colour_map}{Dataframe mapping \code{add_clust_col} to colour values to use
in mapping.}

\item{add_colour_col}{Character name of column in \code{colour_map} that has the
colour values.}
}
\value{
Dataframe with the same columns as \code{eco_desc} but with added rows for
each row of \code{add_eco}
}
\description{
These are usually 'landcover' descriptions.
}
