% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cut_cov}
\alias{cut_cov}
\title{Cover class threshold values}
\format{
A data frame with 5 rows and 2 variables:
\describe{
\item{cov_thresh}{Numeric percentage cover threshold}
\item{pfc}{Text description of projected foliage cover}
...
}
}
\usage{
cut_cov
}
\description{
A lookup table to convert from numeric percentage cover to cover classes.
}
\keyword{datasets}
