% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cut_ht}
\alias{cut_ht}
\title{Height class threshold values}
\format{
A data frame with 9 rows and 1 variables:
\describe{
\item{ht_thresh}{Numeric height (metres) thresholds}
...
}
}
\usage{
cut_ht
}
\description{
A lookup table to convert numeric height to height classes.
}
\keyword{datasets}
