% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_spp_per.R
\name{eco_spp_per}
\alias{eco_spp_per}
\title{Create a dataframe of taxa percent cover for each cluster}
\usage{
eco_spp_per(
  clust_df,
  clust_col = "cluster",
  taxa_df,
  site_col = "cell",
  taxa_col = "taxa",
  cov_col = "cover",
  lustr = NULL
)
}
\arguments{
\item{clust_df}{Dataframe with column indicating cluster membership.}

\item{clust_col}{Name of column in clustdf with cluster membership.}

\item{taxa_df}{Dataframe with columns clustcol and taxacol.}

\item{site_col}{Name of column with 'sites'.}

\item{taxa_col}{Name of column with taxa.}

\item{cov_col}{Name of column with (numeric) cover values.}

\item{lustr}{Dataframe of structural information. Needs columns 'str' and
'storey': structure and storey, respectively.}
}
\value{
Dataframe with columns clustcol, taxacol, str, storey,
}
\description{
Optionally including structural information associated with each taxa.
}
