% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{env}
\alias{env}
\title{Lookup for environmental data}
\format{
A data frame with 55 rows and 19 variables:
\describe{
\item{group}{Broad type of environmental layer.}
\item{category}{Another broad grouping of environmental layer.}
\item{provider}{Who made this environmental layer?}
\item{data_name}{The level at which a reference applies.}
\item{prod}{Three letter code identifying the product.}
\item{band}{Name to any bands in multi-band \code{prod}.}
\item{layer}{Generic name given to that band in file names
derived from \code{prod}.}
\item{process}{What process was used to summarise \code{prod}.}
\item{season_type}{One of; year: whole year, each: split by season or NA}
\item{year_min}{Minimum year available for \code{data_name} and \code{prod}.}
\item{transform}{Numeric. Divide by \code{transform} to convert to \code{units}
from numbers in \code{layer}.}
\item{units}{Units for number in \code{layer} / \code{transform}.}
\item{desc}{Description of the data in \code{prod}.}
\item{detail}{Other relevant details.}
\item{indicates}{Why is the data in \code{layer} useful in predicting ecosystem
distribution?}
\item{link}{Link to webpage for \code{data_name}.}
\item{DOI}{Digital object identifier for \code{data_name}.}
...
}
}
\usage{
env
}
\description{
Lookup for environmental data
}
\keyword{datasets}
