% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lucover}
\alias{lucover}
\title{Lookup for cover codes}
\format{
A data frame with 7 rows and 12 variables:
\describe{
\item{cover_code}{Modified Braun-Blanquet scale}
\item{cover_description}{Text description of cover_code}
\item{cover_desc}{Short text description of cover_code}
\item{cover_seq}{}
\item{cover_1}{1 to 7}
\item{cover_2}{0.01 to 0.75}
\item{cover_3}{0.01 to 5. Sparsely present = 0.5}
\item{cover_4}{0.01 to 5. Sparsely present = 0.1}
\item{pa}{presence/absence. All 1}
\item{cover_mid}{Mid point of scale. 0.01 to 0.875}
\item{cover_max}{Maximum end of scale. 0.01 to 1}
\item{cover_min}{Minimum end of scale. 0 to 0.75}
...
}
}
\usage{
lucover
}
\description{
A lookup table to convert modified Braun-Blanquet scale
\insertRef{RN4265}{envEcosystems} to various numeric scales.
}
\keyword{datasets}
