% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{luenv}
\alias{luenv}
\title{Lookup for environmental data}
\format{
A data frame with 65 rows and 9 variables:
\describe{
\item{data_name}{Character. Source of \code{data_name}}
\item{file}{Character. File name for environmental data}
\item{layer}{Name to use for \code{file}}
\item{desc}{Description of the data in \code{layer}}
\item{group}{Broad groupings of \code{layer} types}
\item{indicator}{Why is the data in \code{layer} useful in predicting ecosystem
distribution?}
\item{transform}{Numeric. Divide by \code{transform} to convert to \code{units}
from numbers in \code{layer}}
\item{units}{Units for number in \code{layer} / \code{transform}}
\item{link}{Link to webpage for \code{data_name}}
\item{link_md}{\code{link} formatted for use in markdown}
...
}
}
\usage{
luenv
}
\description{
Lookup for environmental data
}
\keyword{datasets}
