% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lulandcover}
\alias{lulandcover}
\title{Lookup for landcover classes and their associated colours}
\format{
A data frame with 16 rows and 5 variables:
\describe{
\item{use_class}{Character. Name of landcover class}
\item{definition}{Character}
\item{veg}{Logical. Does the landcover class represent a native ecosystem?}
\item{lc_col}{Character. What colour to map \code{use_class} in land cover map?}
\item{lc_col_veg}{Character. What colour to map \code{use_class} in vegetation map?}
\item{lc_num}{Integer. Useful in sorting}
...
}
}
\usage{
lulandcover
}
\description{
Lookup for landcover classes and their associated colours
}
\keyword{datasets}
