% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lulifeform}
\alias{lulifeform}
\title{Lookup for lifeform codes}
\format{
A data frame with 23 rows and 6 variables:
\describe{
\item{sort}{Sort order for the codes. Essentially from tallest to shortest.}
\item{lifeform}{The code as documented in \insertRef{RN192}{envEcosystems}}
\item{storey}{Over, mid or under storey}
\item{ht}{Height in metres}
\item{str}{Structure as Trees, Mallees, Shrubs, Mat plants, Hummock
grasses, Grasses, Forbs, Sedges, Mistletoe, Ferns or Groundcover}
\item{description}{Text description of lifeform}
...
}
}
\usage{
lulifeform
}
\description{
A lookup table for various lifeform codes, particularly those used in the
\href{https://www.environment.sa.gov.au/topics/Science/Information_data/Biological_databases_of_South_Australia}{Biological Databases of South Australia}
and documented in \emph{Guide to a Native Vegetation Survey}
\insertRef{RN192}{envEcosystems}.
}
\keyword{datasets}
