% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lulifespan}
\alias{lulifespan}
\title{Lookup for lifespan classes}
\format{
A data frame with 4 rows and 2 variables:
\describe{
\item{lifespan}{Character. Lifespan code}
\item{desc}{Character. Lifespan description}
...
}
}
\usage{
lulifespan
}
\description{
Lookup for lifespan classes
}
\keyword{datasets}
