% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eco_desc.R
\name{make_eco_desc}
\alias{make_eco_desc}
\title{Generate a structural description from cluster_taxa_per and the clustering}
\usage{
make_eco_desc(taxa, str, taxonomy)
}
\arguments{
\item{taxa}{Dataframe of taxa information for each ecosystem. Output from
\code{make_eco_taxa_per()}.}

\item{str}{Dataframe of structural information for each ecosystem. Ouput
from \code{make_eco_str_per()}.}

\item{taxonomy}{Dataframe with ind column indicating indigenous status.}
}
\value{
Dataframe of full ecosystem floristic * structural descriptions.
Indcludes text suitable for display in .md
}
\description{
Generate a structural description from cluster_taxa_per and the clustering
}
