% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eco_env_imp.R
\name{make_eco_env}
\alias{make_eco_env}
\title{Make a dataframe summarising environmental variables per cluster.}
\usage{
make_eco_env(
  clust_df,
  imp_df = NULL,
  clust_col = "cluster",
  env_col = "env",
  env_cols
)
}
\arguments{
\item{clust_df}{Dataframe with cluster membership and environmental data.}

\item{imp_df}{Importance from random forest.}

\item{clust_col}{Character. Name of column in \code{clust_df} with cluster
membership.}

\item{env_col}{Character. Name of column in \code{imp_df} with}

\item{env_cols}{Character. Name of columns in \code{clust_df} with environmental
data.}
}
\value{
Dataframe with \code{clust_col}, \code{env_col} and summary columns including
mean, median, standard deviation and (optionally) importance from random
forest.
}
\description{
Optionally include the importance object from random forest.
}
