% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_str_per.R
\name{make_str_per}
\alias{make_str_per}
\title{Create a dataframe of taxa percent cover for each cluster}
\usage{
make_str_per(
  clust_df,
  clust_col = "cluster",
  taxa_df,
  site_col = "cell",
  cov_col = "cover",
  lustr
)
}
\arguments{
\item{clust_df}{Dataframe with column indicating cluster membership.}

\item{clust_col}{Name of column in clustdf with cluster membership.}

\item{taxa_df}{Dataframe with columns site_col, lifeform, and cov_col}

\item{site_col}{Name of column in taxadf with 'sites'.}

\item{cov_col}{Name of column in taxa_df with numeric cover values.}

\item{lustr}{Dataframe of structural information. Needs columns 'lifeform',
'str' and 'storey': structure and storey, respectively. See lulifeform.}
}
\value{
Dataframe with structural information per cluster
}
\description{
Optionally including structural information associated with each taxa.
}
