% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa_sf}
\alias{sa_sf}
\title{Lookup for broad structural formations}
\format{
A data frame with 9 rows and 3 variables:
\describe{
\item{sf}{text structural formation}
\item{colour}{base colour to use in maps/figures}
\item{order}{sort order (roughly lowest to highest)}
...
}
}
\usage{
sa_sf
}
\description{
Lookup for broad structural formations
}
\keyword{datasets}
