% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa_vsf}
\alias{sa_vsf}
\title{Lookup for South Australian vegetation structural formations}
\format{
A data frame with 112 rows and 5 variables:
\describe{
\item{storey}{Ordered factor: over < mid < ground < hummock < fern < mistletoe < mat}
\item{str}{Structure as Trees, Mallees, Shrubs, Mat plants, Hummock
grasses, Grasses, Forbs, Sedges, Mistletoe, Ferns or Groundcover}
\item{cov_class}{Factor cover class, generated by cut(cov_thresh, breaks =
c(unique(cov_thresh),5)}
\item{ht_class}{Factor height class, generated by cut(ht_thresh, breaks =
c(unique(ht_thresh),0)}
\item{sa_vsf}{South Australian vegetation structural formations, as defined
in \emph{Guide to a Native Vegetation Survey} \insertRef{RN192}{envEcosystems}}
...
}
}
\usage{
sa_vsf
}
\description{
As documented in \emph{Guide to a Native Vegetation Survey}
\insertRef{RN192}{envEcosystems}.
}
\keyword{datasets}
