% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_eco_col.R
\name{set_eco_col}
\alias{set_eco_col}
\title{Set colours for individual ecosystems, using gradations within ecotype}
\usage{
set_eco_col(colour, x, y)
}
\arguments{
\item{colour}{Colour for ecotype.}

\item{x}{As in this is ecosystem \code{x} of \code{y} in this ecotype.}

\item{y}{There are \code{y} ecosystems in total for this ecotype.}
}
\value{
hex colour
}
\description{
Set colours for individual ecosystems, using gradations within ecotype
}
