% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_spp_per.R
\name{spp_per}
\alias{spp_per}
\title{Create a dataframe of taxa percent cover for each cluster}
\usage{
spp_per(
  clustdf,
  clustcol = "cluster",
  taxadf,
  sitecol = "cell",
  taxacol = "taxa",
  covcol = "cover",
  lustr = NULL
)
}
\arguments{
\item{clustdf}{Dataframe with column indicating cluster membership.}

\item{clustcol}{Name of column in clustdf with cluster membership.}

\item{taxadf}{Dataframe with columns clustcol and taxacol.}

\item{sitecol}{Name of column in taxadf with 'sites'.}

\item{taxacol}{Name of column in taxadf with taxa.}

\item{lustr}{Dataframe of structural information. Needs columns 'str' and
'storey': structure and storey, respectively.}
}
\value{
Dataframe with columns clustcol, taxacol, str, storey,
}
\description{
Optionally including structural information associated with each taxa.
}
