% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_str_per.R
\name{str_per}
\alias{str_per}
\title{Create a dataframe of taxa percent cover for each cluster}
\usage{
str_per(
  clustdf,
  clustcol = "cluster",
  taxadf,
  sitecol = "cell",
  taxacol = "taxa",
  covcol = "cover",
  lustr
)
}
\arguments{
\item{clustdf}{Dataframe with column indicating cluster membership.}

\item{clustcol}{Name of column in clustdf with cluster membership.}

\item{taxadf}{Dataframe with columns clustcol and taxacol.}

\item{sitecol}{Name of column in taxadf with 'sites'.}

\item{lustr}{Dataframe of structural information. Needs columns 'lifeform',
'str' and 'storey': structure and storey, respectively.}
}
\value{
Dataframe with structural information per cluster
}
\description{
Optionally including structural information associated with each taxa.
}
