% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxa_samphire}
\alias{taxa_samphire}
\title{Vector of samphire taxa}
\format{
A character vector of samphire taxa
...
}
\usage{
taxa_samphire
}
\description{
Adapted from Specht \insertRef{RN175}{envEcosystems}. Used to determine
'samphire' structural formation when these taxa have a combined cover above
a threshold.
}
\keyword{datasets}
