% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxa_wetland}
\alias{taxa_wetland}
\title{Vector of wetland taxa}
\format{
A character vector of wetland taxa
...
}
\usage{
taxa_wetland
}
\description{
Adapted from Specht \insertRef{RN175}{envEcosystems}. Used to determine
'wetland' structural formation when these taxa have a combined cover above
a threshold.
}
\keyword{datasets}
