% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eco_desc.R
\name{add_landcover_desc}
\alias{add_landcover_desc}
\title{Add 'extra' ecosystem (usually landcover) descriptions to ecosystem descriptions}
\usage{
add_landcover_desc(
  eco_desc,
  add_eco,
  clust_col = "cluster",
  add_name = "Landcover",
  colour_map = NULL
)
}
\arguments{
\item{eco_desc}{Dataframe of ecosystem descriptions.}

\item{add_eco}{Dataframe of ecosystems to add.}

\item{clust_col}{Character name of column with cluster membership.}
}
\value{
Dataframe with the same columns as \code{eco_desc} but with added rows for
each row of \code{add_eco}
}
\description{
Add 'extra' ecosystem (usually landcover) descriptions to ecosystem descriptions
}
