% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_desc.R
\name{eco_desc}
\alias{eco_desc}
\title{Generate a structural description from cluster_spp_per and the clustering}
\usage{
eco_desc(strper, sppper, taxonomy)
}
\arguments{
\item{strper}{Dataframe of structural information for each ecosystem. Ouput
from eco_str_per().}

\item{sppper}{Dataframe of taxa information for each ecosystem. Output from
eco_spp_per().}

\item{taxonomy}{Dataframe with ind column indicating indigenous status.}
}
\value{
Dataframe of full ecosystem floristic * structural descriptions.
Indcludes text suitable for display in .md
}
\description{
Generate a structural description from cluster_spp_per and the clustering
}
