% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eco_desc.R
\name{make_eco_desc}
\alias{make_eco_desc}
\title{Title}
\usage{
make_eco_desc(
  bio_df,
  clust_df,
  clust_col = "cluster",
  context,
  cov_col = "use_cover",
  lustr,
  taxonomy,
  use_prop_thresh,
  type_colours = NULL,
  add_eco = NULL,
  add_clust_col = "use_class",
  add_colours = NULL
)
}
\arguments{
\item{bio_df}{Dataframe with biological information.}

\item{clust_df}{Dataframe with cluster membership and join columns to \code{bio_df}.}

\item{clust_col}{Character name of column in \code{clust_df} that defines clusters.}

\item{context}{Character name of columns in \code{bio_df} that define the context.}

\item{cov_col}{Charcter name of column in \code{bio_df} that contain numeric cover
values.}

\item{lustr}{Dataframe containing structural information.}

\item{taxonomy}{Dataframe containing indigenous status of taxa in \code{bio_df}}

\item{use_prop_thresh}{Numeric. Threshold (proportion) for taxa to include in
description. Taxa that occur in more than \code{use_prop_thresh} proportion of
sites in the cluster will be included in the description.}

\item{type_colours}{Colours to use for \code{ecotype}, if available. If no colours
are provided, viridis::viridis is used.}

\item{add_eco}{Dataframe of additional 'ecosystems' to add. These are usually
land cover 'ecosystems' such as, say, cropping.}

\item{add_clust_col}{Character name of column in \code{add_eco} that contains
the additional clusters.}

\item{add_colours}{Dataframe of colours to use for}
}
\value{

}
\description{
Title
}
