% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eco_desc.R
\name{make_eco_desc}
\alias{make_eco_desc}
\title{Make a description for an ecosystem}
\usage{
make_eco_desc(
  bio_df,
  clust_df,
  context,
  clust_col = "cluster",
  taxa_col = "taxa",
  cov_col = "cover_adj",
  ind_abu_col = "use_cover",
  str_col = "lifeform",
  lustr,
  taxonomy,
  use_prop_thresh,
  use_p_val = 0.05,
  n_ind_max = 3,
  ind_val_iter = 3000
)
}
\arguments{
\item{bio_df}{Dataframe containing the site and taxa data in long format.}

\item{clust_df}{Dataframe with \code{context} column(s) and a column with cluster
membership for that context. Optional if \code{clust_col} appears in bio_df.}

\item{context}{Character. Name(s) of column(s) that define the context.}

\item{clust_col}{Character. Name of column containing cluster membership.}

\item{taxa_col}{Character. Name of column containing the taxa names.}

\item{cov_col}{Character. Name of column containing numeric abundance data (
usually 'cover' for plants).}

\item{ind_abu_col}{Character. Name of column containing numeric abundance
data for use in indicator analysis (\code{labdsv::indval()}). This can be the same
as \code{cov_col} but at times a different measure of abundance is helpful here.}

\item{str_col}{Character name of column in \code{bio_df} containing lifeform (or
structural) information.}

\item{lustr}{Dataframe containing lifeform (structural) information.}

\item{taxonomy}{Dataframe containing indigenous status of taxa in \code{bio_df}}

\item{use_prop_thresh}{Numeric. Threshold (proportion) for taxa to include in
description. Taxa that occur in more than \code{use_prop_thresh} proportion of
sites in the cluster will be included in the description.}

\item{use_p_val}{Numeric 0 to 1. The p-value to use to accept a taxa as an
indicator for an ecosystem.}

\item{n_ind_max}{Maximum number of taxa to list as indicators.}

\item{ind_val_iter}{Passed to the \code{...} argument of
\code{envCluster::make_ind_val_df()}, and then into the \code{numitr} argument of
\code{labdsv::indval()}.}
}
\description{
Make a description for an ecosystem
}
