% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eco_desc.R
\name{make_eco_desc}
\alias{make_eco_desc}
\title{Make a description for an ecosystem}
\usage{
make_eco_desc(
  bio_df,
  clust_df,
  bio_wide = NULL,
  bio_ind = NULL,
  context,
  clust_col = "cluster",
  cov_col = "use_cover",
  taxa_col = "taxa",
  str_col = "lifeform",
  ind_col = "ind",
  lustr,
  taxonomy,
  use_prop_thresh,
  use_p_val = 0.05
)
}
\arguments{
\item{bio_df}{Dataframe with biological information.}

\item{clust_df}{Dataframe with cluster membership and join columns to \code{bio_df}.}

\item{bio_wide}{Dataframe of taxa by sites.}

\item{bio_ind}{Dataframe of indicator taxa per class. Output from
\code{envCluster::make_ind_val_df}. Will be created if not supplied.}

\item{context}{Character name of columns in \code{bio_df} that define the context.}

\item{clust_col}{Character name of column in \code{clust_df} that defines clusters.}

\item{cov_col}{Character name of column in \code{bio_df} that contain numeric cover
values.}

\item{taxa_col}{Character name of column with taxa.}

\item{lustr}{Dataframe containing structural information.}

\item{taxonomy}{Dataframe containing indigenous status of taxa in \code{bio_df}}

\item{use_prop_thresh}{Numeric. Threshold (proportion) for taxa to include in
description. Taxa that occur in more than \code{use_prop_thresh} proportion of
sites in the cluster will be included in the description.}

\item{use_p_val}{Numeric 0 to 1. The p-value to use to accept a taxa as an
indicator for an ecosystem.}
}
\description{
Make a description for an ecosystem
}
