% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eco_str_per.R
\name{make_eco_str_per}
\alias{make_eco_str_per}
\title{Create a dataframe of structural information for each cluster}
\usage{
make_eco_str_per(
  bio_clust_df,
  context,
  clust_col = "cluster",
  cov_col = "cover_adj",
  ht_col = "use_height",
  str_col = "lifeform",
  lustr = envEcosystems::lulifeform
)
}
\arguments{
\item{bio_clust_df}{Long format dataframe with: \code{context}(s) defining bins;
cluster membership in \code{clust_col}; and taxa information (\code{cov_col}, \code{ht_col},
\code{str_col}).}

\item{context}{Character. Name of column(s) in \code{bio_clust_df} defining the
context.}

\item{clust_col}{Character. Name of column in \code{bio_clust_df} containing
cluster membership.}

\item{cov_col}{Character. Name of column in \code{bio_clust_df} containing
(numeric, 0 to 1) cover values.}

\item{ht_col}{Character. Name of column in \code{bio_clust_df} containing height
information.}

\item{str_col}{Character. Name of column in \code{bio_clust_df} and \code{lustr}
containing lifeform information.}

\item{lustr}{Dataframe of structural information. Needs column 'str'
(structure) plus \code{str_col}}
}
\value{
Tibble with one row per unique value in \code{clust_col}
}
\description{
Create a dataframe of structural information for each cluster
}
