% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eco_taxa_per.R
\name{make_eco_taxa_per}
\alias{make_eco_taxa_per}
\title{Create a dataframe of taxa percent cover for each cluster}
\usage{
make_eco_taxa_per(
  clust_df,
  clust_col = "cluster",
  bio_df,
  context,
  taxa_col = "taxa",
  cov_col = "cover",
  lustr = NULL
)
}
\arguments{
\item{clust_df}{Dataframe with column indicating cluster membership.}

\item{clust_col}{Name of column in clustdf with cluster membership.}

\item{bio_df}{Dataframe with taxa records. Linked to clust_df by context.}

\item{context}{Name of columns defining the context.}

\item{taxa_col}{Name of column with taxa.}

\item{cov_col}{Name of column with (numeric) cover values.}

\item{lustr}{Dataframe of structural information. Needs columns 'str' and
'storey': structure and storey, respectively.}
}
\value{
Dataframe with columns clustcol, taxacol, str, storey,
}
\description{
Optionally including structural information associated with each taxa.
}
